Split response code from server
